requirements.txt
pytest>=8.3.2
allure-pytest>=2.13.5
selene>=2.0.0rc9
python-dotenv~=1.0.1
selenium~=4.25.0
